Rename app.rs to data.rs
