+++ src/lib.rs
@@ -1,2 +1,2 @@
-pub mod app;
+pub mod data;
 pub mod helper;
